Add an OpenSearch descriptor.
This allows supported browsers to offer to add PBL Links as a
search engine. With this in place, Chrome users should be able
to search PBL Links quickly by typing "pbl<tab>" in the search bar,
and Firefox users will get the "Add PBL Links" option in the search
box to add it as a search engine.